refactor: Split cargo::metadata module
